+++ CHANGELOG.md
@@ -2,8 +2,15 @@
 
 ## UNRELEASED
 
-* Added Swift 3 support (Xcode 8 beta 2) while retaining Swift 2 support.  
+* Added Swift 3 support (Xcode 8) while retaining Swift 2 support.  
   [@phatblat](https://github.com/phatblat), [#20](https://github.com/AliSoftware/Reusable/pull/20)
+
+## 2.5.0
+
+* Added the possibility for `NibOwnerLoadable` confirming custom views to pass an existing instance as `owner`
+  (used as the File's Owner) in case one already exists. This is especially useful to implement `init(coder:)`
+  to load the content of the XIB as subviews of `self` after initialization. See `MyCustomWidget.swift` for an example.  
+  [@AliSoftware](https://github.com/AliSoftware)
 
 ## 2.4.0